refactor: prevent search route collision
