+++ api/contacts/search/index.ts
@@ -1,5 +1,6 @@
-const getAirtableContext = require("../../lib/airtableBase");
-const { createSearchHandler } = require("../../lib/airtableSearch");
+// Moved to prevent route collision with [id].ts in Next.js
+const getAirtableContext = require("../../../lib/airtableBase");
+const { createSearchHandler } = require("../../../lib/airtableSearch");
 
 const apiContactsSearchHandler = async (req: any, res: any) => {
   const { TABLES } = getAirtableContext();
