+++ api/snapshots/search/index.ts
@@ -1,6 +1,7 @@
-const getAirtableContext = require("../../lib/airtableBase");
-const { createSearchHandler } = require("../../lib/airtableSearch");
-const { getFieldMap } = require("../../lib/resolveFieldMap");
+// Moved to prevent route collision with [id].ts in Next.js
+const getAirtableContext = require("../../../lib/airtableBase");
+const { createSearchHandler } = require("../../../lib/airtableSearch");
+const { getFieldMap } = require("../../../lib/resolveFieldMap");
 
 const apiSnapshotsSearchHandler = async (req: any, res: any) => {
   const { TABLES } = getAirtableContext();
